+++ package.json
@@ -20,6 +20,7 @@
   "homepage": "https://github.com/German969/books-challenge#readme",
   "dependencies": {
     "bcryptjs": "^2.4.3",
+    "cookie-parser": "^1.4.6",
     "dotenv": "^16.4.5",
     "ejs": "^3.1.6",
     "express": "^4.17.2",
