+++ src/views/partials/header.ejs
@@ -9,6 +9,9 @@
       <li><a href="/authors" class="nav-link px-2 link-dark">Authors</a></li>
       <li><a href="/books/search" class="nav-link px-2 link-dark">Search</a></li>
     </ul>
+    
+<!-- ------------------------------------------------------------------------------------ -->
+<!--     si en la variable locals hay "usuarioLogueado" como true, se mostrara solo el logout, por caso contrario mostrara login y signup -->
 <% if (locals.usuarioLogueado) { %>
   <div>
   <a href="/logout" class="btn btn-outline-primary me-2">Logout</a>
